Move reliability test into its own file
The other messy integration test breaks it quite incidentally...
